WR-2 Afficher la list de tous les employees

+++ src/main/java/Services/DepartementService.java
@@ -19,7 +19,7 @@ public class DepartementService implements DepartementServiceInterface {
     }
 
     @Override
-    public Departement getDepartementById(Long id) {
+    public Departement getDepartementById(int id) {
         return departementRepository.findById(id);
     }
