+++ src/main/java/Services/ServicesInterfaces/DepartementServiceInterface.java
@@ -5,7 +5,7 @@ import Models.Departement;
 import java.util.List;
 public interface DepartementServiceInterface {
     void addDepartement(Departement departement);
-    Departement getDepartementById(Long id);
+    Departement getDepartementById(int id);
     List<Departement> getAllDepartements();
     void updateDepartement(Departement departement);
     void removeDepartement(int id);
