+++ src/main/java/Services/ServicesInterfaces/EmployeeServiceInterface.java
@@ -5,7 +5,7 @@ import Models.Employee;
 import java.util.List;
 public interface EmployeeServiceInterface {
     void addEmployee(Employee employee);
-    Employee getEmployeeById(Long id);
+    Employee getEmployeeById(int id);
     List<Employee> getAllEmployees();
     void updateEmployee(Employee employee);
     void removeEmployee(int id);
